Added new case when Spring beans are being overriden unnoticed
This case is when we have two Spring beans with the same name in two
different configuration files and then one of those config files
is importing the other. Then Spring overrides beans and it is very
highly possible that we will be using in our code beans that are not
correct and that are different specialization that we expect.
This is scary shit though as it will not explode like the previous example
with @Qualifier annotation. I might not even notice it it might run
for a very long time !!